2_ai_vision/analyze_image/requirements.txt
# azure-cognitiveservices-vision-computervision
# msrest
azure-ai-vision-imageanalysis